refactor: rename score edge folder to edge
prepping to add another edge component
